chore(aggregated-assets): move platform layer behind shims (LIVE-35227)
The 9 hooks, the 3 cache selectors, discovery and currency selection move to
@features/platform-aggregated-assets. Eleven one-line re-export shims stay at the
old dada-client paths, so no consumer's production imports change.

The selector files get no shim: nothing outside the package imported them, so
they stay internal. discovery and currencySelection are exported because Global
Search, the Stocks section and the asset selectors call them directly.

Tests move with their subjects, colocated per the DDD guideline. That is not
cosmetic: the hooks used to read the endpoint from a different module than
parseError and mergeAssetsDataPages, so a factory mock of state-manager/api left
those alone. Now one module supplies all three and the factories need
jest.requireActual spread in, or the hooks crash on a missing
mergeAssetsDataPages. Assertions are untouched.

@shared/env is mocked in jest.setup.js because the features CI job does not
install libs/env, so @shared/env cannot resolve @ledgerhq/live-env and any test
importing the domain barrel fails to load.

selectorUtils' cache scan names the bucket it indexes, removing the implicit-any
desktop reports on develop. The cross-query scan is unchanged; reworking it is
LIVE-35301.
