+++ features/platform/aggregated-assets/src/hooks/useAssetsData.test.ts
@@ -3,10 +3,11 @@
  */
 
 import { renderHook } from "@testing-library/react";
-import { useAssetsData } from "../useAssetsData";
-import { useGetAssetsDataInfiniteQuery } from "../../state-manager/api";
+import { useAssetsData } from "./useAssetsData";
+import { useGetAssetsDataInfiniteQuery } from "@domain/api-aggregated-assets";
 
-jest.mock("../../state-manager/api", () => ({
+jest.mock("@domain/api-aggregated-assets", () => ({
+  ...jest.requireActual("@domain/api-aggregated-assets"),
   useGetAssetsDataInfiniteQuery: jest.fn(),
 }));
