+++ features/platform/aggregated-assets/src/hooks/useChunkedAssetsData.test.ts
@@ -3,10 +3,11 @@
  */
 
 import { renderHook } from "@testing-library/react";
-import { useChunkedAssetsData } from "../useChunkedAssetsData";
-import { useGetChunkedAssetsDataQuery } from "../../state-manager/api";
+import { useChunkedAssetsData } from "./useChunkedAssetsData";
+import { useGetChunkedAssetsDataQuery } from "@domain/api-aggregated-assets";
 
-jest.mock("../../state-manager/api", () => ({
+jest.mock("@domain/api-aggregated-assets", () => ({
+  ...jest.requireActual("@domain/api-aggregated-assets"),
   useGetChunkedAssetsDataQuery: jest.fn(),
 }));
