+++ features/platform/aggregated-assets/src/hooks/useMarketByCurrencies.ts
@@ -1,0 +1,28 @@
+import { useSelector } from "react-redux";
+import { CryptoOrTokenCurrency } from "@domain/entity-currency";
+import isEqual from "lodash/isEqual";
+import { selectMarketByCurrency } from "../selectors/marketSelectors";
+import { ApiState } from "../selectors/selectorUtils";
+
+export type MarketByCurrencies = Partial<
+  Record<string, { price: number; priceChangePercentage24h: number }>
+>;
+
+export const useMarketByCurrencies = (currencies: CryptoOrTokenCurrency[]) => {
+  return useSelector((state: ApiState) => {
+    const marketByCurrencies: MarketByCurrencies = {};
+    for (const currency of currencies) {
+      const currencyMarket = selectMarketByCurrency(state, currency.id);
+      if (
+        currencyMarket?.price !== undefined &&
+        currencyMarket?.priceChangePercentage24h !== undefined
+      ) {
+        marketByCurrencies[currency.id] = {
+          price: currencyMarket.price,
+          priceChangePercentage24h: Math.round(currencyMarket.priceChangePercentage24h * 100) / 100,
+        };
+      }
+    }
+    return marketByCurrencies;
+  }, isEqual);
+};
