+++ features/platform/aggregated-assets/src/index.ts
@@ -1,1 +1,20 @@
-export {};
+export { useAssetsData } from "./hooks/useAssetsData";
+export { useAssetData } from "./hooks/useAssetData";
+export { useChunkedAssetsData } from "./hooks/useChunkedAssetsData";
+export { useStocksData } from "./hooks/useStocksData";
+export { useStockAssetIds } from "./hooks/useStockAssetIds";
+export { useStablecoinTickers } from "./hooks/useStablecoinTickers";
+export { useInterestRatesByCurrencies } from "./hooks/useInterestRatesByCurrencies";
+export { useMarketByCurrencies } from "./hooks/useMarketByCurrencies";
+export { useLazyLedgerCurrency } from "./hooks/useLazyLedgerCurrency";
+
+export type { InterestRatesByCurrencies } from "./hooks/useInterestRatesByCurrencies";
+
+export {
+  selectTopStocks,
+  selectTopAssetsByCategory,
+  type StockSuggestion,
+  type CategorizedDiscoveryAsset,
+  type CategorizedDiscoveryAssets,
+} from "./discovery";
+export { selectCurrency, selectCurrencyForMetaId } from "./currencySelection";
