+++ libs/ledger-live-common/src/dada-client/hooks/useMarketByCurrencies.ts
@@ -1,28 +1,1 @@
-import { useSelector } from "react-redux";
-import { CryptoOrTokenCurrency } from "@domain/entity-currency";
-import isEqual from "lodash/isEqual";
-import { selectMarketByCurrency } from "../entities/marketSelectors";
-import { ApiState } from "../entities/selectorUtils";
-
-export type MarketByCurrencies = Partial<
-  Record<string, { price: number; priceChangePercentage24h: number }>
->;
-
-export const useMarketByCurrencies = (currencies: CryptoOrTokenCurrency[]) => {
-  return useSelector((state: ApiState) => {
-    const marketByCurrencies: MarketByCurrencies = {};
-    for (const currency of currencies) {
-      const currencyMarket = selectMarketByCurrency(state, currency.id);
-      if (
-        currencyMarket?.price !== undefined &&
-        currencyMarket?.priceChangePercentage24h !== undefined
-      ) {
-        marketByCurrencies[currency.id] = {
-          price: currencyMarket.price,
-          priceChangePercentage24h: Math.round(currencyMarket.priceChangePercentage24h * 100) / 100,
-        };
-      }
-    }
-    return marketByCurrencies;
-  }, isEqual);
-};
+export { useMarketByCurrencies } from "@features/platform-aggregated-assets";
