+++ libs/ledger-live-common/src/dada-client/hooks/useStablecoinTickers.ts
@@ -1,21 +1,1 @@
-import { useMemo } from "react";
-import { useGetAssetsByCategoryQuery } from "../state-manager/api";
-import { AssetCategory } from "../state-manager/types";
-
-const emptySet = new Set<string>();
-
-export function useStablecoinTickers(product: "llm" | "lld", version: string, skip?: boolean) {
-  const { data, isLoading, isError } = useGetAssetsByCategoryQuery(
-    {
-      category: AssetCategory.Stablecoins,
-      product,
-      version,
-    },
-    { skip },
-  );
-  const tickers = useMemo(
-    () => (data ? new Set(data.map(t => t.toUpperCase())) : emptySet),
-    [data],
-  );
-  return { tickers, isLoading, isError };
-}
+export { useStablecoinTickers } from "@features/platform-aggregated-assets";
